feat: comment out everything collect_coins related, tests are wip
