+++ pallets/creditcoin/src/ocw/tasks/verify_transfer.rs
@@ -1,9 +1,14 @@
 use ethabi::{Function, Param, ParamType, StateMutability, Token};
 use ethereum_types::U64;
-use frame_support::ensure;
+use frame_support::{ensure, RuntimeDebug};
 use frame_system::pallet_prelude::BlockNumberFor;
 use sp_core::U256;
 use sp_runtime::traits::UniqueSaturatedFrom;
+use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
+use scale_info::TypeInfo;
+use sp_core::H160;
+use hex_literal::hex;
+
 #[cfg(not(feature = "std"))]
 use sp_std::prelude::*;
 
@@ -16,6 +21,47 @@ use crate::{
 	Blockchain, Config, ExternalAddress, ExternalAmount, ExternalTxId, Id, OrderId, Transfer,
 	TransferKind, UnverifiedTransfer,
 };
+
+#[derive(Clone, Encode, Decode, Eq, PartialEq, RuntimeDebug, TypeInfo, MaxEncodedLen)]
+pub struct DeployedContract {
+	pub address: sp_core::H160,
+	pub chain: Blockchain,
+}
+
+impl DeployedContract {
+	const DEFAULT_CHAIN: Blockchain = Blockchain::Ethereum;
+}
+
+impl Default for DeployedContract {
+	fn default() -> Self {
+		let contract_chain: Blockchain = DeployedContract::DEFAULT_CHAIN;
+		let contract_address: H160 =
+			sp_core::H160(hex!("a3EE21C306A700E682AbCdfe9BaA6A08F3820419"));
+		Self { address: contract_address, chain: contract_chain }
+	}
+}
+
+impl DeployedContract {
+	///exchange has been deprecated, use burn instead
+	fn burn_vested_cc_abi() -> Function {
+		#[allow(deprecated)]
+		Function {
+			name: "burn".into(),
+			inputs: vec![Param {
+				name: "value".into(),
+				kind: ParamType::Uint(256),
+				internal_type: None,
+			}],
+			outputs: vec![Param {
+				name: "success".into(),
+				kind: ParamType::Bool,
+				internal_type: None,
+			}],
+			constant: Some(false),
+			state_mutability: StateMutability::NonPayable,
+		}
+	}
+}
 
 pub(crate) fn ethless_transfer_function_abi() -> Function {
 	#[allow(deprecated)]
